# Replace C implementation of OS Random engine with Python one that just calls os.urandom