This is alternate fix for #2007 that moves file descriptor shenanigans (and C code) out of cryptography entirely.  It is intended to supersede #2035 .