@glyph 2015-06-28
The last commit addresses @alex's last bit of feedback, so I think this is good to go now?