@alex 2015-06-28
Besides that comment this LGTM, @reaperhulk what do you think?

Now that this is more easily worked with (because python), we should file some follow up issues to write tests that stub out `os.urandom` and verify that the results are used correctly.