@alex 2015-06-29
@reaperhulk 8192 is a pretty big read for `/dev/urandom`, if you've got the script handy, could you try with 16 byte reads?