@reaperhulk 2015-06-29
100,000 16 byte reads takes ~0.6-0.7s with the C engine code, ~0.9-1.0s with this PR.