@reaperhulk 2015-06-29
Here's an even more minimal script that attempts to solely measure the cost of the `RAND_bytes` call (my previous code allocated a new buffer and also pulled out the data after each call)

```python
from cryptography.hazmat.backends.openssl import backend

buf = backend._ffi.new("char[16]")
for _ in range(100000):
    res = backend._lib.RAND_bytes(buf, 16)
    assert res == 1
```

On my machine running `time python testrand.py` (best run of 5)
### This PR
```
real	0m0.848s
user	0m0.567s
sys	0m0.276s
```
Variance between 0.848s and 0.860s

### Master
```
real	0m0.535s
user	0m0.326s
sys	0m0.205s
```
Variance between 0.535s and 0.554s


I don't necessarily think this is going to be a serious problem, just something to be aware of. It could conceivably negatively affect TLS handshake performance? PyOpenSSL won't be affected though because it doesn't activate this engine.