@glyph 2015-06-29
I cranked up the iteration count by another 10x, switched to `xrange` instead of `range`, and measured with PyPy.

best of 3:

# This PR
```
real	0m2.193s
user	0m0.731s
sys	0m1.463s
```
Variance between 2.292 and 2.193

# Master

```
real	0m1.857s
user	0m0.407s
sys	0m1.447s
```
Variance between 1.857 and 1.945

It looks to me like we're pretty clearly just measuring function call / buffer copy overhead for your Python runtime here, and pypy does commensurately better.  How much random data does one TLS handshake require?  My suspicion is that this amount of additional overhead (on pypy, approximately 3µs/call - and it is per *call*, not per byte, as the overhead all but disappears with fewer calls generating the same number of bytes) will verge on undetectable.  Just ballpark-wise, http://www.semicomplete.com/blog/geekery/ssl-latency.html (from 2010) puts TLS handshake latency at 322000µs, so we'd have to be calling `RAND_bytes` 1000 times in the course of the handshake to cause a 1% degradation to performance.