@alex 2015-06-29
Yeah, I think the performance implications of this are totally 100% fine :-)