@glyph 2015-06-30
That last commit should address [@reaperhulk's feedback](https://github.com/pyca/cryptography/pull/2073#issuecomment-116356191). Anything else?