@glyph 2015-06-30
> The patched test is failing on libressl.

whyyyyyyyy

OK. Any tips on how to set up a libressl build environment that I can test with?