@glyph 2015-06-30
For those following along at home, I did a `brew install libressl` and then this helpful shell function to switch OpenSSL's:
```bash
function kegger () {
    for kegtoadd in "$@"; do
        local kegpath="/usr/local/opt/$kegtoadd";
        export LDFLAGS="-L${kegpath}/lib $LDFLAGS";
        export CPPFLAGS="-I${kegpath}include $CPPFLAGS";
        export PATH="${kegpath}/bin:$PATH";
    done;
}
```