@glyph 2015-06-30
Regarding the build failures:

- the travis failure is a spurious issue with installing dependencies on the docs builder.
- the codecov failure is a result of LibreSSL not being part of the test matrix; I added that code as a result of @reaperhulk's feedback, so I'm assuming LibreSSL *should* be in that matrix?