@reaperhulk 2015-07-01
LGTM, awesome work @glyph. We should (probably as a separate PR) update our docs around the OS random engine to note explicitly that it uses the Python `os.urandom` implementation (and what that entails) along with a note that LibreSSL patches random itself.