@alex 2015-07-01
Merge?

On Tue, Jun 30, 2015 at 8:18 PM, Paul Kehrer <notifications@github.com>
wrote:

> LGTM, awesome work @glyph <https://github.com/glyph>. We should (probably
> as a separate PR) update our docs around the OS random engine to note
> explicitly that it uses the Python os.urandom implementation (and what
> that entails) along with a note that LibreSSL patches random itself.
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/pyca/cryptography/pull/2073#issuecomment-117377257>.
>



-- 
"I disapprove of what you say, but I will defend to the death your right to
say it." -- Evelyn Beatrice Hall (summarizing Voltaire)
"The people's good is the highest law." -- Cicero
GPG Key fingerprint: 125F 5C67 DFE9 4084